pyproject.toml
[tool.poetry]
name = "binday"
version = "0.1.0"
description = ""
authors = ["Chehan Ratnasiri <chehan.rat@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyserial = "^3.4"
flask = "^1.1.2"
python-dotenv = "^0.13.0"
flask-migrate = "^2.5.3"
Flask-SQLAlchemy = "^2.4.1"
flask-login = "^0.5.0"
Flask-WTF = "^0.14.3"
flask-bootstrap = "^3.3.7"
wtforms = {extras = ["email"], version = "^2.3.1"}
WTForms-Components = "^0.10.4"
colour = "^0.1.5"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.1"
isort = "^4.3.21"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
